Moved account_creation_view into auth section while also shortening it's code, adding that code into account_creation_control. Email verification now works properly!
