+++ lib/src/auth/account_creation_view.dart
@@ -1,0 +1,81 @@
+// ignore_for_file: prefer_const_constructors
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter_svg/svg.dart';
+import 'package:microbial_source_tracking/src/home/home_view.dart';
+import 'package:firebase_storage/firebase_storage.dart';
+import 'package:microbial_source_tracking/src/login/login_view.dart';
+
+import '../widgets/account_creation_controller.dart';
+import '../widgets/login_controller.dart';
+
+//RegisterAccount is the name of this widget,
+//refer to RegisterAccount for routing purposes
+class RegisterView extends StatelessWidget {
+  const RegisterView({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: const Color.fromARGB(255, 233, 248, 255),
+      body: SingleChildScrollView(
+        child: SafeArea(
+          child: Container(
+              width: double.infinity,
+              padding: const EdgeInsets.all(8.0),
+              child: Column(
+                children: [
+                  const SizedBox(height: 10),
+                  Container(
+                      alignment: Alignment.center,
+                      child: SvgPicture.asset('assets/images/glwa_logo.svg')),
+                  const SizedBox(height: 50),
+                  const Text(
+                    'Create a new account',
+                    style: TextStyle(
+                        color: Colors.black,
+                        fontSize: 20,
+                        fontWeight: FontWeight.w500),
+                  ),
+                  const SizedBox(height: 15),
+
+                  //Register widget
+                  Register(),
+                  const SizedBox(height: 20),
+                ],
+              )),
+        ),
+      ),
+      bottomNavigationBar: Container(
+        height: 50,
+        alignment: Alignment.topCenter,
+        child: Row(mainAxisAlignment: MainAxisAlignment.center, children: [
+          const Text('Already have an account? ',
+              style: TextStyle(fontSize: 20)),
+          InkWell(
+            onTap: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (context) => const LoginView(),
+                ),
+              );
+            },
+            child: const Text(
+              'Return to login',
+              style: TextStyle(
+                  color: Colors.blue,
+                  fontWeight: FontWeight.bold,
+                  fontSize: 20),
+            ),
+          )
+        ]),
+      ),
+    );
+  }
+}
